remove nesting
